workign subset + infinite mode

+++ src/results.js
@@ -10,8 +10,11 @@ for (let i = 0; i < data.length; i++) {
 }
 
 document.querySelector("#clear-btn").onclick = () => {
-    localStorage.clear()
-    window.location.href = '/'
+    const answer = confirm("Delete local data? (server data is retained in the public dataset)");
+    if (answer) {
+      localStorage.clear()
+      window.location.href = '/'
+    }
 }
 
 
@@ -19,8 +22,10 @@ const resultsSummaryElement = document.querySelector("#results-summary")
 if (localStorage.getItem('results')) {
     const results = JSON.parse(localStorage.getItem('results'))
     const guesses = results.guesses
+    const correct_num = results.guesses.filter(g => g.correct).length
+    const accuracy = Math.round(((correct_num) / (guesses.length)) * 100)
 
-    resultsSummaryElement.textContent = `Total answered: ${guesses.length}. Correct: ${results.guesses.filter(g => g.correct).length}`
+    resultsSummaryElement.textContent = `Accuracy: ${accuracy}%. Total answered: ${guesses.length}. Correct: ${correct_num}`
     const resultsTable = document.querySelector("#results-table")
     for (let i = 0; i < guesses.length; i++) {
         const currentGuess = guesses[i]
@@ -28,6 +33,7 @@ if (localStorage.getItem('results')) {
         const { job_title, guess } = currentGuess
         const correctAnswer = jobsMap[job_title]
         const symbol = (correctAnswer == guess) ? '✅' : '❌'
+        currentGuess.answer = correctAnswer
 
         let className = 'blue'
         if (correctAnswer == 'republican') className = 'red'
@@ -38,6 +44,17 @@ if (localStorage.getItem('results')) {
         <td>${guess}</td>
         <td class=${className}>${correctAnswer}</td>`
         resultsTable.appendChild(trNode)
+    }
+
+    document.querySelector("#download-data-btn").onclick = () => {
+      const csvData = jsonToCSV(guesses);
+      const blob = new Blob([csvData], { type: 'text/csv' });
+      const url = window.URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.setAttribute('href', url);
+      a.setAttribute('download', 'shape-of-america-user-data.csv');
+      a.click();
+      window.URL.revokeObjectURL(url);
     }
 } else {
     resultsSummaryElement.textContent = 'No results found!'
@@ -84,3 +101,21 @@ document.querySelector("#results-table-container").querySelectorAll('th')
     window.column_toggled[idx] = !window.column_toggled[idx]
     sortTable(idx, window.column_toggled[idx])
 })));
+
+
+
+
+
+//// Helpers 
+
+function jsonToCSV(jsonData) {
+  const keys = Object.keys(jsonData[0]);
+  const csvRows = [keys.join(",")];
+
+  jsonData.forEach(row => {
+      const values = keys.map(key => `"${row[key]}"`);
+      csvRows.push(values.join(","));
+  });
+
+  return csvRows.join("\n");
+}
